电网视频生成助手/pyproject.toml
[project]
name = "grid-video-agent"
version = "0.1.0"
description = "A starter agent that turns grid industry news into short-video production assets."
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115",
  "httpx>=0.27",
  "imageio>=2.34",
  "numpy>=1.26",
  "pillow>=10.0",
  "pydantic>=2.8",
  "python-dotenv>=1.0",
  "uvicorn>=0.30",
  "volcengine>=1.0.218",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "src*"]
